Improve cartesian product logic and UI
Updated the cartesian product logic to be based on iterating using
indices from the longest row rather than the first. This makes
it more flexible and we don't need to insist the first row is
somehow "special". Also updated the UI to better reflect what is
going on under the hood -- the user can now see which fields are
being used to generate the cartesian product based on their vertical
alignment.